requirements.txt
Django==4.2.19
djangorestframework==3.15.2
django-cors-headers==4.6.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b109
numpy==1.26.4
pandas
scikit-learn
catboost==1.2.7
feedparser==6.0.11
apscheduler==3.11.0
pytz==2024.2
requests==2.32.3
python-dateutil==2.9.0
yandex-cloud-ml-sdk==0.3.0
redis
ipywidgets
ipython
traitlets
psycopg2-binary
sqlalchemy
